Seperated deck manager and edit deck page components into two seperated folders for better readability
