+++ frontend/src/components/deckmanagerpagecomponents/display/owneddeckdisplaycomponent.tsx
