+++ frontend/src/components/deckmanagerpagecomponents/types/homepagecomponentprops.ts
