+++ frontend/src/components/editdeckpagecomponents/carddropzones/SideDeckCardsZone.tsx
@@ -2,9 +2,9 @@ import { useDroppable } from "@dnd-kit/core"
 import { useState } from "react"
 import { faMinus, faPlus, faTrash } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { handleDecreaseCardOwnedAmount, handleDeleteCardClick, handleIncreaseCardOwnedAmount } from "./EditDeckCardButtons"
-import { SideDeckProps } from "../types/editdeckcomptypes"
-import SubGridListViewComponent from "../draganddropitems/subgridlistviewcomp"
+import { handleDecreaseCardOwnedAmount, handleDeleteCardClick, handleIncreaseCardOwnedAmount } from "../../editdeckpagecomponents/buttons/EditDeckCardButtons"
+import { SideDeckProps } from "../types/carddropzonetypes"
+import SubGridListViewComponent from "../../editdeckpagecomponents/draganddropitems/subgridlistviewcomp"
 
 const SideDeckCardZone = ({ sidedeckprops }: SideDeckProps) => {
   const {
