+++ frontend/src/components/editdeckpagecomponents/draganddropitems/CollectionGalleryViewItem.tsx
