+++ frontend/src/components/editdeckpagecomponents/draganddropitems/CollectionListViewItem.tsx
