+++ frontend/src/components/editdeckpagecomponents/draganddropitems/SearchCardGalleryViewItem.tsx
