+++ frontend/src/components/editdeckpagecomponents/draganddropitems/SearchCardListViewItem.tsx
